tests/submission/W3C/domhtml/HTMLDocument20.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/dynamic-markup-insertion.html#dom-document-open" rel="help">
<link rel="author" title="Curt Arnold">
</head>
<body id="TEST-BODY"> 
<p>
Replaces the current document with a valid HTML document using HTMLDocument.open, writeln and close.
</p>

<iframe id='iframe' src='/support/html/sample.html'>
</iframe>


<script>
var domtest = async_test("Replaces the current document with a valid HTML document using HTMLDocument.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var doc;
      var docElem;
      var title;
      doc = window.document.getElementById('iframe').contentWindow.document;
      assert_equals(doc.title, "Sample");
      doc.open();
      doc.writeln("<html>");
      doc.writeln("<body>");
      doc.writeln("<title>Replacement</title>");
      doc.writeln("</body>");
      doc.writeln("<p>");
      doc.writeln("Hello, World.");
      doc.writeln("</p>");
      doc.writeln("</body>");
      doc.writeln("</html>");
      doc.close();
      assert_equals(doc.title, "Replacement");
      
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
